Add files via upload
Sentiment Analysis and Recommender System on the YELP Dataset. 
In this project, we are going to use a SVM model to predict the ratings of the restaurants from the reviews of the users. What is interesting is to use the coefficients associated with each word to infer what is appreciated and what is not for a particular restaurant. The idea is to use all the reviews of a restaurant to predict the associated ratings. Once this is done, we can exctract the coefficients of the words and see their magnitude and sign. If "Price" has a negative coefficient, it means that that restaurant is probably overpriced, if "staff" is positive, it means that clients likes the people working there. 
Moreover, here you can find two examples of recommender systems, on naive and one more sophisticated, using matrix factorization.